+++ etc/config.xml
@@ -86,7 +86,6 @@ tr.shipping.totals td.amount span.price,
                 <catalog_ingestion_instant_enabled>1</catalog_ingestion_instant_enabled>
                 <catalog_ingestion_instant_async_enabled>0</catalog_ingestion_instant_async_enabled>
                 <catalog_ingestion_instant_events>product_stock_status</catalog_ingestion_instant_events>
-                <catalog_ingestion_system_configuration_update_request_enabled>1</catalog_ingestion_system_configuration_update_request_enabled>
                 <pdp_frequency_selector>0</pdp_frequency_selector>
             </boltpay>
         </payment>
